docs: update font preview with blockquote note
