+++ website/package.json
@@ -20,8 +20,8 @@
     "@chakra-ui/theme-tools": "^1.1.7",
     "@emotion/react": "^11.4.0",
     "@emotion/styled": "^11.3.0",
-    "@fontsource/inter": "^4.4.1",
-    "@fontsource/raleway": "^4.4.0",
+    "@fontsource/inter": "^4.4.2",
+    "@fontsource/raleway": "^4.4.2",
     "@mdx-js/loader": "^1.6.22",
     "@mdx-js/react": "^1.6.22",
     "@next/mdx": "^10.2.3",
@@ -40,12 +40,12 @@
     "remark-slug": "^6.0.0"
   },
   "devDependencies": {
-    "@types/node": "^15.6.1",
-    "@types/react": "^17.0.8",
-    "@types/react-dom": "^17.0.5",
+    "@types/node": "^15.12.0",
+    "@types/react": "^17.0.9",
+    "@types/react-dom": "^17.0.6",
     "@types/react-instantsearch-dom": "^6.10.0",
-    "@typescript-eslint/eslint-plugin": "^4.25.0",
-    "@typescript-eslint/parser": "^4.25.0",
+    "@typescript-eslint/eslint-plugin": "^4.26.0",
+    "@typescript-eslint/parser": "^4.26.0",
     "eslint": "^7.27.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-import": "^2.23.4",
@@ -55,7 +55,7 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "next-remote-watch": "^1.0.0",
-    "next-sitemap": "^1.6.95",
+    "next-sitemap": "^1.6.99",
     "prettier": "^2.3.0",
     "typescript": "^4.3.2"
   },
