+++ website/src/components/MdxComponents.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable react/display-name */
 import {
-  Alert,
   Box,
   Code as InlineCode,
   Divider,
@@ -16,27 +15,11 @@ import {
   Thead,
   Tr,
   UnorderedList,
-  useColorModeValue,
 } from "@chakra-ui/react";
 
+import { BlockQuote } from "./Blockquote";
 import { Code } from "./Code";
 import { NextChakraLink } from "./NextChakraLink";
-
-const BlockQuote = (props) => {
-  const borderColor = useColorModeValue("gray.300", "gray.700");
-  const bgColor = useColorModeValue("gray.100", "gray.800");
-  return (
-    <Alert
-      as="blockquote"
-      variant="left-accent"
-      background={bgColor}
-      borderInlineStartColor={borderColor}
-      py={0}
-      mt={2}
-      {...props}
-    />
-  );
-};
 
 const CustomMdxComponents = {
   // Typography
